Fixed Double Hints in SqlInjection Advanced Challenge
