+++ webgoat-lessons/sql-injection/src/main/resources/i18n/WebGoatLabels.properties
@@ -5,6 +5,7 @@ sql.advanced.title=SQL Injection (advanced)
 
 sql-injection.error=<span class='feedback-negative'>Sorry, this solution is not correct. Try again!</span>
 NoResultsMatched=<span class='feedback-negative'>No results matched. Try Again.</span>
+ResultsButNotTom=<span class='feedback-negative'>Try To login as Tom!</span>
 
 sql-injection.2.success=<span class='feedback-positive'>You have succeeded!</span>
 sql-injection.2.failed=<span class='feedback-negative'>Something went wrong! You got no results, check your SQL Statement and the table above.</span>
